+++ source/codegen/metadata/nifgen/CHANGES.md
@@ -31,6 +31,10 @@ The following functions were tagged with their corresponding CNAME tag
 
 Added value_twist attribute to the coefficientsArray parameter of GetFirFilterCoefficients function.
 
+Added 'grpc_type' tag to these APIs since they use custom types : 
+- CreateWaveformComplexF64
+- WriteComplexBinary16Waveform
+
 The following functions removed as they are obsolete:
 - ConfigureGain
 - ConfigureRefClockFrequency